requirements.txt
asyncpraw
asyncprawcore
discord.py
ffmpeg
Flask
gTTS
PyNaCl
requests
urllib3
Werkzeug
yarl
youtube-dl
SQLAlchemy